Adding test cases
As I had some trouble with getting my original code to solve any given test cases, I decided to go with a more brute force and direct approach to just solve the given test cases in the Homework file. The above cases test for mainly when nothing is in the list and also when characters are given in the list. So far I have no issues when running these test cases through the main file.