Adding test cases
This is another more brute force and direct method of going about the solution. These are non-number and empty list test cases. The main function works perfectly for non-nested lists and purely number lists but I realized that it needed a little help with dealing with nested lists, which I tested next.