The main function
So what I did here was similar to what I did for #1. This is a recursive solution. The first clause deals with the base case, when there is only one element left in the list, that is your result. The second clause deals with the recursion step. It grabs the first two items of the list and performs a recursive call replacing those two items with a new term L + Sum1. This however only works with non-nested lists that contain both numbers and non-numbers